Move type transmission checks to common package
